+++ cpp/include/ZMat.h
@@ -1,5 +1,9 @@
 #ifndef ZMAT_H_INCLUDED
 #define ZMAT_H_INCLUDED
+/**
+ * The pre-defined header.
+ * @author ZHG <CyberZHG@gmail.com>
+ */
 
 #define ZMAT_BEGIN_NAMESPACE namespace zmat {
 #define ZMAT_END_NAMESPACE };